requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.3
defusedxml==0.5.0
Django==1.11.1
django-allauth==0.32.0
django-authtools==1.5.0
django-filter==1.0.4
djangorestframework==3.6.2
idna==2.5
Markdown==2.6.8
oauthlib==2.0.2
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
python3-openid==3.1.0
pytz==2017.2
requests==2.17.3
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.21.1
